+++ eras/conway/impl/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.17.0.0
 
-* Changed `ConwayWdrlNotDelegatedToDRep` to wrap `NonEmpty`
+* Changed `ConwayWdrlNotDelegatedToDRep` to wrap `NonEmpty KeyHash`
 * Add `showGovActionType`, `acceptedByEveryone`
 * Added `unRatifySignal`
 * Added lenses: